Revert "Merge remote-tracking branch 'upstream/bugfix-2.0.x' into Printer"
This reverts commit 36a489dcf882eaef87027090e526082c0839a571, reversing
changes made to 88d7b0ef2ab3f5b19fbb9b7dfeae6edbe8be3fd8.

+++ Marlin/src/lcd/extui/lib/dgus/DGUSDisplay.cpp
@@ -30,6 +30,10 @@
   #error "More than 2 hotends not implemented on the Display UI design."
 #endif
 
+#include "DGUSDisplay.h"
+#include "DGUSVPVariable.h"
+#include "DGUSDisplayDef.h"
+
 #include "../../ui_api.h"
 
 #include "../../../../MarlinCore.h"
@@ -43,10 +47,6 @@
 #if ENABLED(POWER_LOSS_RECOVERY)
   #include "../../../../feature/powerloss.h"
 #endif
-
-#include "DGUSDisplay.h"
-#include "DGUSVPVariable.h"
-#include "DGUSDisplayDef.h"
 
 // Preamble... 2 Bytes, usually 0x5A 0xA5, but configurable
 constexpr uint8_t DGUS_HEADER1 = 0x5A;
@@ -440,6 +440,7 @@ void DGUSScreenVariableHandler::DGUSLCD_SendHeaterStatusToDisplay(DGUS_VP_Variab
       snprintf_P(tmpfilename, VP_SD_FileName_LEN, PSTR("%s%c"), filelist.filename(), filelist.isDir() ? '/' : 0);
     DGUSLCD_SendStringToDisplay(var);
   }
+
 
   void DGUSScreenVariableHandler::SDCardInserted() {
     top_file = 0;
@@ -854,7 +855,7 @@ void DGUSScreenVariableHandler::HandleStepPerMMExtruderChanged(DGUS_VP_Variable 
 void DGUSScreenVariableHandler::HandleProbeOffsetZChanged(DGUS_VP_Variable &var, void *val_ptr) {
   DEBUG_ECHOLNPGM("HandleProbeOffsetZChanged");
 
-  const float offset = float(int16_t(swap16(*(uint16_t*)val_ptr))) / 100.0f;
+  const float offset = float(swap16(*(uint16_t*)val_ptr)) / 100.0f;
   ExtUI::setZOffset_mm(offset);
   ScreenHandler.skipVP = var.VP; // don't overwrite value the next update time as the display might autoincrement in parallel
   return;
